consolidate misc libanari headers into a single frontend/ directory
