+++ code_gen/templates/ArrayObjects.cpp
@@ -1,6 +1,6 @@
 // This file was generated by $generator from $template
 
-#include "anari/type_utility.h"
+#include "anari/anari_cpp.hpp"
 #include "$prefixSpecializations.h"
 
 #include <cstdlib>
