+++ code_gen/templates/FrameObject.cpp
@@ -1,7 +1,7 @@
 // This file was generated by $generator from $template
 
 #include "$prefixSpecializations.h"
-#include "anari/type_utility.h"
+#include "anari/anari_cpp.hpp"
 
 #include <cstdlib>
 #include <cstring>
