+++ code_gen/templates/Parameter.h
@@ -2,8 +2,7 @@
 
 #pragma once
 
-#include <anari/type_utility.h>
-#include <anari/anari_cpp/Traits.h>
+#include <anari/anari_cpp.hpp>
 #include <cstring>
 #include <new>
