+++ examples/generated_device_frontend/device/src/TreeArrayObjects.cpp
@@ -3,7 +3,7 @@
 
 // This file was generated by generate_device_frontend.py from ArrayObjects.cpp
 
-#include "anari/type_utility.h"
+#include "anari/anari_cpp.hpp"
 #include "TreeSpecializations.h"
 
 #include <cstdlib>
