+++ examples/generated_device_frontend/device/src/TreeFrameObject.cpp
@@ -4,7 +4,7 @@
 // This file was generated by generate_device_frontend.py from FrameObject.cpp
 
 #include "TreeSpecializations.h"
-#include "anari/type_utility.h"
+#include "anari/anari_cpp.hpp"
 
 #include <cstdlib>
 #include <cstring>
