+++ libs/anari/anari_sdk_version.h.in
