+++ libs/anari/include/anari/frontend/anari_enums.h
