+++ libs/anari/include/anari/frontend/type_utility.h
